requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
django-redis==5.2.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
orjson==3.8.14
psycopg2-binary==2.9.6
gunicorn==20.1.0
pydantic==1.10.8
factory_boy==3.2.1
Faker==18.9.0
openpyxl==3.1.2
requests==2.31.0
